Findcourses US & UK Content Fields
(Same proposal I left on 2.0) Hi guys, entering a bit of uncharted territory here but I hope this makes sense - we're phasing out usage of custom content fields on the UK and US sites and it would be great to have this reflected in the documentation for all future and current XML clients. 

It would also be much appreciated if we could add 'Training Course Content' (UK) and 'Training Content' (US) as default content fields as they are the only ones currently missing from the list of defaults already set up. 

I've made an update to try and reflect this, not sure if it's in the right place/or if I've done this whole thing correctly but just give me a shout if there's a better method. Cheers!