content-fields
Had a shot at creating a file this time - here are the content fields we're using for each site. All of these are our site 'defaults' - anything not listed here won't become active upon import hence the need of having something like this to guide any new and current XML clients towards. 

We don't have any clients for educations.com using XML for now but I thought it would be worth adding just in case. If this works, it would also be great to have a link to it in the main 'course' documentation page (if possible). Curious to hear any feedback/ideas for improvement! Thank you so much!